requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
celery[redis]==5.2.7
flower==1.2.0

# dev пакеты
django-debug-toolbar
django-extensions
rich
ipython